move logic in separate hook
